migrated day 7, finished day 13 part 1

+++ src/utils/vec2.kt
@@ -2,17 +2,17 @@ package utils
 
 import kotlin.math.sqrt
 
-typealias Vec2 = Pair<Int,Int>
+typealias Vec2 = Pair<Long,Long>
 
-val Vec2.x: Int
+val Vec2.x: Long
     get() = this.first
 
-val Vec2.y: Int
+val Vec2.y: Long
     get() = this.second
 
 operator fun Vec2.plus(b: Vec2) = Pair(x + b.x, y + b.y)
 operator fun Vec2.minus(b: Vec2) = Pair(x - b.x, y - b.y)
-operator fun Vec2.times(b: Int) = Pair(x*b, y*b)
+operator fun Vec2.times(b: Long) = Pair(x*b, y*b)
 fun Vec2.mag() = sqrt((x*x + y*y).toDouble())
 
 fun Vec2.up(): Vec2 {
